Moved CouponUploadListener.java
